Updated Sign-in and Sign-up layouts
I updated the layouts of the sign-in and sign-up screens to closely match the mockup. To do so I had to change the create account button to a textview which meant I needed to update the LoginActivity files as well. The only changes to the SignUpActivity were from id updates.