Added the create chatroom fragment
Added the create chatroom fragment which allows the user to put the name of a new chatroom. It stores that chatroom name and the user id in realtime db.

I also updated some of the layouts and  some other small changes. (edit to ProfileFragment to allow transition to updateProfile, change to MainActivity to add the Create Chatroom option to the slide drawer.)