Updated Storage Layout, and methods of storage
Likes are stored as map instead of list, active users are now in the chatroom instead of off by themselves. Also a naming convention issue with backslash '/'.

Please feel free to test and make sure nothing has broken. Didn't have time to work on the new active users display (or the new image links that can be referenced from anywhere or whatever the professor was saying.)