Added share location feature
Added share location feature

It is pretty basic, the user that shares also gets a notification that a location was shared but at least it is implemented.

Feel free to make changes (the same button will also remove the shared location.)

Just now thinking about it, it might have meant to continuously share the location. If that is the case, perhaps pushpdeep can work on that while he is implementing it in the other screens.