Added the server file for node js code for apis
Added the server file for node js code for apis and I added comments and some code for parameter collection to call the api I created in the server.js file.

1. I created the simeple text to text translation api endpoint.
It takes a body of
{
    "from":"en",
    "to":["it"],
    "message":"Message to translate"
}
and will return response containing the transalted message.

2. I added the code to collect the parameters (i.e. if English is selected I added the code that should select "en" and so on and so forth.

@valmray I was not entirely sure how to write a request in Kotlin so I did not do that yet.

After setting up this endpoint I think we should be in a good spot come Wednesday. I will try to work more on the APIs tomorrow evening.

P.S. I removed the subscriptionKey from the server.js file so that it is not exposed on GitHub. So when testing you would need to collect that from the shared file we have in private (AND MAKE SURE TO REMOVE IT BEFORE PUSHING TO GITHUB, I almost forgot to...)