Added code to the app that will recognize speech into text
Added code to the app that will recognize speech into text (I did not use azure speech to text it was not working for me).

Still need to add the code that will call the translation API if the user selects to translate their speech. Also would probably be best to have a popup dialog before the speech recognition takes place so we know ahead of time if it should be translated or not.

The speech sdk stuff is still in the project because I had already added it so did not see the point in removing it. Can also go back at any point to remove them.