Added logic to speech translation page that will actually speak in other languages
Added logic to speech translation page that will actually speak in other languages.

The logic is only on the speech translation page.

I had to hardcode the possible languages to the ones that are able to be spoken/synthesized.

@valmray we will need to update the translate functionality on the cardview as well. We will get rid of the Language collection API call, as the languages are hardcoded in the Strings file with the codes, codes_with_region, and voices.

I can help with the conversion if you have some time tonight or after 5 tomorrow.